move CPUTopology and ThreadAffinity to toplevel
